smart block concatination
